chore: Shift all the types from index.ts to types.ts
