+++ packages/hooks/src/index.ts
@@ -1,32 +1,9 @@
 /**
  * Internal dependencies
  */
-import type { Hooks as internalHooks } from './createHooks';
 import createHooks from './createHooks';
 
-export type Callback = ( ...args: any[] ) => any;
-
-export type Handler = {
-	callback: Callback;
-	namespace: string;
-	priority: number;
-};
-
-export type Hook = {
-	handlers: Handler[];
-	runs: number;
-};
-
-export type Current = {
-	name: string;
-	currentIndex: number;
-};
-
-export type Store = Record< string, Hook > & { __current: Set< Current > };
-
-export type StoreKey = 'actions' | 'filters';
-
-export type Hooks = internalHooks;
+export * from './types';
 
 export const defaultHooks = createHooks();
